Suppressing mne warnings. Renamed logging module to plogging to avoid naming conflicts.
