+++ niprov/__init__.py
@@ -1,6 +1,6 @@
 from niprov.discovery import discover
 from niprov.inspection import inspect
-from niprov.logging import log
+from niprov.plogging import log
 from niprov.recording import record
 from niprov.adding import add
 from niprov.renaming import renameDicoms
